requirements.txt
alembic==1.7.7
asttokens==2.0.5
autopep8==1.6.0
Babel==2.9.1
backcall==0.2.0
blinker==1.4
click==8.1.2
colorama==0.4.4
decorator==5.1.1
executing==0.8.3
Flask==2.1.1
Flask-Admin==1.6.0
Flask-BabelEx==0.9.4
Flask-Login==0.6.0
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
Flask-Principal==0.4.0
Flask-Security==3.0.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
greenlet==1.1.2
importlib-metadata==4.11.3
importlib-resources==5.6.0
ipython==8.2.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.1
Mako==1.2.0
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
parso==0.8.3
passlib==1.7.4
pickleshare==0.7.5
prompt-toolkit==3.0.29
psycopg2==2.9.3
pure-eval==0.2.2
pycodestyle==2.8.0
Pygments==2.11.2
python-dotenv==0.20.0
pytz==2022.1
six==1.16.0
speaklater==1.3
SQLAlchemy==1.4.34
stack-data==0.2.0
toml==0.10.2
traitlets==5.1.1
wcwidth==0.2.5
Werkzeug==2.1.1
WTForms==3.0.1
zipp==3.8.0
